requirements.txt
gitpython==3.1.44
pyyaml==6.0.2
